libfdt warning fixes

+++ scripts/dtc/libfdt/fdt.c
@@ -134,13 +134,13 @@ const void *fdt_offset_ptr(const void *fdt, int offset, unsigned int len)
 	unsigned absoffset = offset + fdt_off_dt_struct(fdt);
 
 	if (fdt_chk_basic())
-		if ((absoffset < offset)
+		if ((absoffset < (unsigned int)offset)
 		    || ((absoffset + len) < absoffset)
 		    || (absoffset + len) > fdt_totalsize(fdt))
 			return NULL;
 
 	if (!fdt_chk_version() || fdt_version(fdt) >= 0x11)
-		if (((offset + len) < offset)
+		if (((offset + len) < (unsigned int)offset)
 		    || ((offset + len) > fdt_size_dt_struct(fdt)))
 			return NULL;
 
@@ -304,7 +304,7 @@ int fdt_move(const void *fdt, void *buf, int bufsize)
 {
 	FDT_RO_PROBE(fdt);
 
-	if (fdt_totalsize(fdt) > bufsize)
+	if (fdt_totalsize(fdt) > (unsigned int)bufsize)
 		return -FDT_ERR_NOSPACE;
 
 	memmove(buf, fdt, fdt_totalsize(fdt));
